+++ Services/ImageSearch/V2/Model/SearchOptionalParam.cs
@@ -38,7 +38,7 @@ namespace HuaweiCloud.SDK.ImageSearch.V2.Model
         /// 对象类目，如给定则不进行对象分类，直接使用该category作为类目。具体类目信息可参见对应的服务类型说明。
         /// </summary>
         [JsonProperty("category", NullValueHandling = NullValueHandling.Ignore)]
-        public decimal? Category { get; set; }
+        public int? Category { get; set; }
 
         /// <summary>
         /// 去重标签名，必须为服务实例custom_tags中已存在的key。 - 如给定则会对该key下相同value的数据进行去重，仅保留得分最高的数据。 - 针对没有设置该标签的数据，会直接过滤。
